+++ src/components/common/CustomButton.jsx
@@ -1,4 +1,4 @@
-import styles from '../styles.module.css'
+import styles from '../../styles.module.css'
 
 const CustomButton = ({ title, handleClick, style }) => {
     return (
